Exercise 14
João Papo-de-Pescador, a good man, bought a microcomputer to control his daily work output. Every time he brings a weight of fish greater than that established by the fishing regulations of the state of São Paulo (50 kilos) he must pay a fine of R$4.00 per excess kilo. João needs you to write a program that reads the weight variable (fish weight) and calculates the excess. Record in the excess variable the number of kilos beyond the limit and in the fine variable the amount of the fine that João must pay. Print program data with appropriate messages.