Exercise 17
Make a Program for a paint store. The program should ask for the size in square meters of the area to be painted. Consider that the paint coverage is 1 liter for every 6 square meters and that the paint is sold in 18-liter cans, which cost R$80.00 or in 3.6-liter gallons, which cost R$25.00 .
Inform the user of the quantities of paint to be purchased and the respective prices in 3 situations:
buy only 18 liter cans;
buy only 3.6 liter gallons;
mix cans and gallons, so that paint waste is less. Add 10% slack and always round the values ​​up, that is, consider full cans.