Exercise 12
Create a program to calculate a payroll, knowing that the deductions are from Income Tax, which depend on the gross salary (as per the table below) and 3% for the Union and that the FGTS corresponds to 11% of the Gross Salary, but it is not discounted (it is the company that deposits it). Net Salary corresponds to Gross Salary minus discounts. The program should ask the user for their hourly rate and the number of hours worked in the month.
IR Discount:
Gross Salary up to 900 (inclusive) - exempt
Gross Salary up to 1500 (inclusive) - 5% discount
Gross Salary up to 2500 (inclusive) - 10% discount
Gross Salary above 2500 - 20% discount Print the information on the screen, arranged as per the example below. In the example, the hour value is 5 and the hour quantity is 220.
        Gross Salary: (5*220): R$ 1100.00
        (-) IR (5%): R$ 55.00  
        (-) INSS (10%): R$ 110.00
        FGTS (11%): R$ 121.00
        Total discounts: R$ 165.00
        Net Salary: R$ 935.00