Exercise 15
Make a Program that asks for the 3 sides of a triangle. The program should inform you whether the values ​​can be a triangle. Indicate, if the sides form a triangle, whether it is: equilateral, isosceles or scalene.
Tips:
Three sides form a triangle when the sum of any two sides is greater than the third;
Equilateral Triangle: three equal sides;
Isosceles Triangle: any two equal sides;
Scalene Triangle: three different sides;