Exercise 20
Make a program to read three partial notes from a student. The program must calculate the average achieved per student and present:
The message "Approved", if the average is greater than or equal to 7, with the respective average achieved;
The message "Failed", if the average is less than 7, with the respective average achieved;
The message "Approved with Distinction", if the average is equal to 10.