Exercise 24
Make a program that asks a person 5 questions about a crime. The questions are:
"Did you call the victim?"
"Were you at the crime scene?"
"Do you live near the victim?"
"Did you owe the victim anything?"
"Have you worked with the victim?" The program should then give a classification of the person's participation in the crime. If the person answers positively to 2 questions, he or she should be classified as "Suspect", between 3 and 4 as "Accomplice" and 5 as "Murderer". Otherwise, he or she will be classified as "Innocent".