Exercise 29
Mr. Manoel Joaquim owns a large store that sells items for R$1.99, with about 10 cash registers. To speed up the calculation of how much each customer should pay, he developed a table that contains the number of items the customer has purchased and the amount of the bill next to it. This way, the cashier only needs to count how many items the customer is taking and look at the price list. You have been hired to develop the program that creates this price list, which will contain the prices of 1 to 50 products, as shown in the example below: Lojas Quase Dois - Price list 1 - R$1.99 2 - R$3.98 ... 50 - R$99.50