Exercise 30
Mr. Manoel Joaquim has just purchased a bakery and intends to implement the price list methodology, which is already a success in his $1.99 store. You have been hired to develop the program that creates a price list for breads, from 1 to 50 loaves, based on the price of the bread informed by the user, as shown in the example below: Price of bread: R$ 0.18 Pão de Ontem Bakery - Price list 1 - R$ 0.18 2 - R$ 0.36 ... 50 - R$ 9.00