Exercise 31
Mr. Manoel Joaquim has expanded his business beyond the 1.99 business and now owns a convenience store.
Write a program that implements a rudimentary cash register. The program should receive an unknown number of
values ​​referring to the prices of the goods. A value of zero should be entered by the operator to indicate the end of the
purchase. The program should then display the total of the purchase and ask for the amount of money that the customer provided, and then calculate and display the amount of change. After this operation, the program should return to the starting point, to record
the next purchase. The output should be as shown in the example below:
Tabajara Stores
Product 1: R$ 2.20
Product 2: R$ 5.80
Product 3: R$ 0
Total: R$ 9.00
Money: R$ 20.00
Change: R$ 11.00