Exercise 37
A gym wants to take a survey of its customers to find out who is the tallest, shortest, fattest and thinnest. To do this, you must create a program that asks each of the gym's customers for their code, height and weight. The data entry should end when the user enters 0 (zero) in the code field. When the program ends, the codes and values ​​of the tallest, shortest, fattest and thinnest customer should also be informed, as well as the average height and weight of the customers.