Exercise 38
An employee of a company receives an annual salary increase: It is known that:
This employee was hired in 1995, with an initial salary of R$ 1,000.00;
In 1996, he received a 1.5% increase on his initial salary;
From 1997 onwards (inclusive), salary increases always correspond to double the percentage of the previous year.
Create a program that determines this employee's current salary. After completing this, modify the program to allow
the user to enter the employee's initial salary.