Exercise 40
Statistics were conducted in five Brazilian cities to collect data on traffic accidents. The following data were obtained:

City code;
Number of passenger vehicles (in 1999);
Number of traffic accidents with victims (in 1999). The aim is to know:
What is the highest and lowest rate of traffic accidents and to which city it belongs;
What is the average number of vehicles in the five cities together;
What is the average number of traffic accidents in cities with less than 2,000 passenger vehicles.