Exercise 43
The menu of a snack bar is as follows:

Specification Code Price
Hot Dog 100 R$ 1.20
Bauru Plain 101 R$ 1.30
Bauru with egg 102 R$ 1.50
Hamburger 103 R$ 1.20
Cheeseburger 104 R$ 1.30
Soft drink 105 R$ 1.00

Make a program that reads the code of the items ordered and the desired quantities. Calculate and display the amount to be paid per item (price * quantity) and the total of the order. Consider that the customer must inform when the order should be
closed.