Exercise 44
In a presidential election, there are four candidates. The votes are reported using a code. The codes used are: 1, 2, 3, 4 - Votes for the respective candidates (you should create a table, for example: 1 - Jose/ 2- João/etc.) 5 - Invalid Vote 6 - Blank Vote Write a program that calculates and displays: The total votes for each candidate; The total number of invalid votes; The total number of blank votes; The percentage of invalid votes over the total number of votes; The percentage of blank votes over the total number of votes. The final set of votes is zero.