Exercise 45
Develop a program to check a student's grade on a test with 10 questions. The program should ask the student the answer to each question and then compare it with the answer sheet to calculate the total number of correct answers and the grade (assign 1 point for each correct answer). After each student uses the system, a question should be asked if another student will use the system. After all students have answered, inform: Highest and Lowest Correct Answers; Total Number of Students who used the system; Average Grades for the Class. Test Answer Sheet: 01 - A 02 - B 03 - C 04 - D 05 - E 06 - E 07 - D 08 - C 09 - B 10 - A After completing this, you could improve the program by allowing the teacher to type the answer sheet for the test before the students use the program.