Exercise 46
In a long jump competition, each athlete is entitled to five jumps. At the end of each athlete's series of jumps, the best and worst results are eliminated. Your result is the average of the three remaining values. You must write a program that receives the name and the five distances achieved by the athlete in his jumps and then informs the average of the jumps according to the description given above (remove the best and worst jumps and then calculate the average). Use a list to store the jumps. The jumps are informed in the order in which they are executed, so they are not ordered. The program must be terminated when the athlete's name is not informed. The program output should be as shown in the example below: Athlete: Rodrigo Curvêllo First Jump: 6.5 m Second Jump: 6.1 m Third Jump: 6.2 m Fourth Jump: 5.4 m Fifth Jump: 5.3 m Best Jump: 6.5 m Worst Jump: 5.3 m Average of other jumps: 5.9 m Final Result: Rodrigo Curvêllo: 5.9 m