Exercise 05
Write a program that converts from 24-hour time to 12-hour time. For example, the program should convert 2:25 PM to 2:25 P.M. The input is two integers. There should be at least two functions: one to do the conversion and one to output. Record the A.M./P.M. information as a value of ‘A’ for A.M. and ‘P’ for P.M. The function that performs the conversions will then have a formal parameter to record whether it is A.M. or P.M. Include a loop that allows the user to repeat this calculation for new input values ​​as many times as desired.