Exercise 09
Craps Game. Write a program to implement a Craps game.
The player rolls a pair of dice, getting a value between 2 and 12.
If, on the first roll, you roll a 7 or 11, you are a "natural" and win.
If you roll a 2, 3, or 12 on the first roll, this is called "craps" and you lose.
If, on the first roll, you rolled a 4, 5, 6, 8, 9, or 10, this is your "point."
Your goal now is to keep rolling the dice until you roll this number again.
You lose, however, if you roll a 7 before rolling this point again.