Exercise 11
Scramble words. Build a function that takes a string as a parameter and returns another string with the scrambled characters.
For example: if the function receives the word python, it might return npthyo, ophtyn, or any other possible combination,
randomly. Standardize in your function that all characters will be returned in uppercase or lowercase,
regardless of how they were typed.