Exercise 13
Magic square. A magic square is one divided into rows and columns, with a number in each position, and in which the sum of the rows, columns, and diagonals is the same. For example, consider a magic square with a side of 3, with numbers from 1 to 9:
8 3 4
1 5 9
6 7 2
Create a function that identifies and displays on the screen all magic squares with the above characteristics. Tip:
Produce all possible combinations and check the sum when completing each square. Using a vector from 1 to 9 seems
to be simpler than using a 3x3 matrix.